change setupserver location
